org cem width
